requirements.txt
asgiref==3.7.2
Django==5.0.1
sqlparse==0.4.4
mysql-connector-python==8.0.30
mysqlclient==2.1.1